Rename sub-module "data_reduction" -> "moments"; also remove masking parts of "Mask_and_Moments" and rename to "Moments"
