--- turbustat/data_reduction/__init__.py
@@ -1,1 +1,0 @@
-from .make_moments import Mask_and_Moments
