+++ turbustat/moments/__init__.py
@@ -1,0 +1,1 @@
+from .make_moments import Moments
